changed the path to cart,contact and about
