+++ customer/cart.php
@@ -2,10 +2,10 @@
 
 session_start();
 
-require_once("config/database.php");
-include("includes/head.php");
-include("customer/includes/navbar.php");
-include("functions/functions.php");
+require_once'../config/database.php';
+include'includes/head.php';
+include'includes/navbar.php';
+//include'functions/functions.php';
 
 ?>
 
@@ -286,30 +286,8 @@ include("functions/functions.php");
         ?>
 
         <?php
-
-        function update_cart(){
-
-          global $con;
-
-          if(isset($_POST['update'])){
-
-              foreach($_POST['remove'] as $remove_id){
-
-
-              $delete_product = "DELETE from cart where p_id='$remove_id'";
-
-              $run_delete = mysqli_query($con,$delete_product);
-
-              if($run_delete){
-                echo "<script>window.open('cart.php','_self')</script>";
-              }
-            }
-          }
-        }
-        // echo @$up_cart = update_cart();
-
+        update_cart();
         ?>
-
         <div class="row">
 
           <div class="col-3"><!-- col-md-3 col-sm-6 Starts -->
