+++ customer/contact.php
@@ -2,7 +2,7 @@
 
 session_start();
 
-require_once("config/database.php");
+require'../config/database.php';
 include("includes/header.php");
 include("functions/functions.php");
 include("includes/main.php");
